# packages/bc-wallet-openapi/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Credential Showcase API
  version: 0.1.1
servers:
  - url: https://virtserver.swaggerhub.com/SphereonInt/showcase/0.1.0/v1
    description: SwaggerHub API Auto Mocking
  - url: https://bcshowcase-api-dev.nborbit.ca/
...
security:
  - OAuth2: [ ]
  - OIDC: [ ]
paths:
...
  /personas:
    get:
      tags:
        - Personas
      summary: List personas
      operationId: listPersonas
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonasResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Personas
      summary: Create a persona
      operationId: createPersona
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonaRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonaResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /personas/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Personas
      summary: Get a specific persona
      operationId: getPersona
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonaResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Personas
      summary: Update a persona
      operationId: updatePersona
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonaRequest'
        required: true
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonaResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Personas
      summary: Delete a persona
      operationId: deletePersona
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: <https://auth.example.com/oauth/authorize>
          tokenUrl: <https://auth.example.com/oauth/token>
          scopes:
            editor: Grants editor access
            admin: Grants admin access
            viewer: Grants viewer access
    OIDC:
      type: openIdConnect
      openIdConnectUrl: <https://auth.example.com/.well-known/openid-configuration>
      description: OpenID Connect security scheme for OAuth2 flows
  schemas:
...
    Asset:
      type: object
      required:
        - id
        - mediaType
        - content
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Unique identifier for the asset in the database
          example: 123e4567-e89b-12d3-a456-426614174469
        mediaType:
          type: string
          description: Type of asset (e.g., image, document, icon)
          example: image/jpeg
        content:
          type: string
          description: Binary data of the asset
          example: base64 encoded binary data
        fileName:
          type: string
          description: Name of the asset file
          example: asset.jpg
        description:
          type: string
          description: Description of the asset
          example: A beautiful image of a cat
        createdAt:
          type: string
          description: Date and time the asset was created
          format: date-time
          example: 2025-03-06T11:30:16.573Z
        updatedAt:
          type: string
          description: Date and time the asset was last updated
          format: date-time
          example: 2025-03-06T11:30:16.573Z
...
    Persona:
      type: object
      required:
        - id
        - name
        - slug
        - role
        - description
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Unique identifier for the persona
          example: 123e4567-e89b-12d3-a456-426614174456
        name:
          type: string
          description: Name of the persona
          example: John Doe
        slug:
          type: string
          description: URL-friendly identifier for the persona
          example: john-doe
        role:
          type: string
          description: Role of the persona in the system
          example: Verifier
        description:
          type: string
          description: Detailed description of the persona
          example: John Doe is a verifier for the system
        headshotImage:
          $ref: '#/components/schemas/Asset'
        bodyImage:
          $ref: '#/components/schemas/Asset'
        hidden:
          type: boolean
          description: Whether the persona is hidden from the user
          example: false
        createdAt:
          type: string
          description: Date and time the persona was created
          format: date-time
          example: 2025-03-06T11:30:16.573Z
        updatedAt:
          type: string
          description: Date and time the persona was last updated
          format: date-time
          example: 2025-03-06T11:30:16.573Z
    PersonaRequest:
      type: object
      required:
        - name
        - role
      properties:
        name:
          type: string
          description: Name of the persona
          example: John Doe
        role:
          type: string
          description: Role of the persona in the system
          example: Verifier
        description:
          type: string
          description: Detailed description of the persona
          example: John Doe is a verifier for the system
        headshotImage:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174456
        bodyImage:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174877
        hidden:
          type: boolean
          description: Whether the persona is hidden from the user
          example: false
    PersonasResponse:
      properties:
        personas:
          type: array
          items:
            $ref: '#/components/schemas/Persona'
    PersonaResponse:
      type: object
      properties:
        persona:
          $ref: '#/components/schemas/Persona'
...
  responses:
    'BadRequest':
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Invalid request data'
    'Unauthorized':
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Invalid authentication credentials'
    'Forbidden':
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Insufficient permissions to access resource'
...
    'InternalServerError':
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'An unexpected error occurred'